Поправил модель
git-svn-id: https://svtz.ru/svn/graphlabs.site/branches/taskresults@337 e8704808-1a2a-453a-8987-66899a109cbe

+++ GraphLabs.Site.Models/Results/ResultModel.cs
@@ -9,7 +9,7 @@ namespace GraphLabs.Site.Models.Results
         public long Id { get; set; }
         public string Mode { get; set; }
         public DateTime StartDateTime { get; set; }
-        public short? Score { get; set; }
+        public int? Score { get; set; }
         public string Status { get; set; }
         public string LabWorkName { get; set; }
         public string LabVariantNumber { get; set; }
